style : style 파일 제거
