+++ src/containers/InformationPage/components/Category.tsx
