+++ src/containers/InformationPage/components/index.ts
@@ -1,4 +1,4 @@
-export { Search } from './search/Search';
-export { FixedButton } from './fixedButton/FixedButton';
-export { Notice } from './notice/Notice';
-export { Category } from './category/Category';
+export { Search } from './Search';
+export { FixedButton } from './FixedButton';
+export { Notice } from './Notice';
+export { Category } from './Category';
